backend_python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.9.0
google-patent-scraper>=1.0.8
beautifulsoup4>=4.12.0
lxml>=5.0.0
